+++ source/application/AMPIC887DriverServer/AMPIC887DataRecorderState.h
@@ -34,12 +34,12 @@ public:
 	  * \param recordTrigger ~ The trigger to initialize the state with.
 	  * \param availableRecordParameters ~ Text describing each of the parameters
 	  * available for recording on the controller.
-	  * \param recordRate ~ The record Rate of the controller.
+	  * \param recordRate ~ The record Rate of the controller in hertz.
 	  * \param recordConfigs ~ The record configs to initialize the state with.
 	  */
 	void initialize(AMGCS2::DataRecordTrigger recordTrigger,
 					const QString& availableRecordParameters,
-					int recordRate,
+					double recordRate,
 					const QHash<int, AMPIC887DataRecorderConfiguration>& recordConfigs);
 
 	/*!
@@ -68,15 +68,15 @@ public:
 	void setRecordConfig(int tableId, AMPIC887DataRecorderConfiguration recordConfig);
 
 	/*!
-	  * The current data recorder record rate, in controller cycles.
+	  * The current data recorder record rate, in hertz.
 	  */
-	int recordRate() const;
+	double recordRate() const;
 
 	/*!
 	  * Sets the controller record rate to the provided value
-	  * \param recordRate ~ The new record rate to set (in controller cycles).
+	  * \param recordRate ~ The new record rate to set (in hertz).
 	  */
-	void setRecordRate(int recordRate);
+	void setRecordRate(double recordRate);
 
 	/*!
 	  * The record configs for all the data tables of the controller.
@@ -105,7 +105,7 @@ protected:
 	AMGCS2::DataRecordTrigger recordTrigger_;
 	QString availableRecordParameters_;
 	QHash<int, AMPIC887DataRecorderConfiguration> recordConfigs_;
-	int recordRate_;
+	double recordRate_;
 
 };
